refactor: modularize dependency and preemption logic under 300-line limit
