+++ aura-core/src/orchestrator/commands/mod.rs
@@ -6,6 +6,7 @@ use tracing::info;
 
 pub(crate) mod add;
 pub(crate) mod config;
+pub(crate) mod dependency;
 pub(crate) mod lifecycle;
 pub(crate) mod retry;
